71_выносит из дата класса всю логику
